Remove WKWebView from automatic monitoring
This is not reliable since WKWebView and WebKit run in a different process.
There is an IPC between the app and WebKit process but some data (i.e.e POST body) are not transmitted to the app process.
This would lead to unexpected behaviour. POST data would be discarded for instance.