Remove webview from automatic monitoring (#4)
Remove WKWebView from automatic monitoring

Automatic monitoring of the WKWebView URLSession is not reliable since the WKWebView and WebKit run in different processes.
There is an IPC between the app and WebKit process but some data (i.e. POST body) are not transmitted to the app process.
This would lead to an unexpected behaviour. POST data would be discarded for instance.